+++ CMakeLists.txt
@@ -33,3 +33,6 @@ if(AA_ENABLE_TESTING)
 	include(CTest)
 	add_subdirectory(tests)
 endif()
+if(MSVC)
+  target_compile_options(${PROJECT_NAME}lib INTERFACE "/Zc:__cplusplus")
+endif()
