+++ tests/CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_minimum_required(VERSION 3.21)
 
-set(AA_PARTS anyany data_parallel_vector variant_swarm)
+set(AA_PARTS anyany data_parallel_vector variant_swarm visit_invoke)
 foreach(PART ${AA_PARTS})
 	add_executable(${PART} ${CMAKE_CURRENT_SOURCE_DIR}/test_${PART}.cpp)
 	target_include_directories(${PART}
@@ -12,7 +12,10 @@ foreach(PART ${AA_PARTS})
 		LINKER_LANGUAGE CXX
 		CMAKE_CXX_STANDARD_REQUIRED ON
 		CXX_STANDARD 20
-	)	
+	)
+	if(MSVC)
+		target_compile_options(${PART} PUBLIC "/Zc:__cplusplus")
+	endif()
 endforeach()
 
 file(GLOB AA_TESTS REGULAR_EXPRESSION ${CMAKE_CURRENT_SOURCE_DIR}/*.cpp)
